Updated unittest to test import of solar_theme. Renamed test module.
