--- tests/test_solar.py
@@ -1,28 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-test_solar
-----------------------------------
-
-Tests for `solar` module.
-"""
-
-import unittest
-
-from solar import solar
-
-
-class TestSolar(unittest.TestCase):
-
-    def setUp(self):
-        pass
-
-    def test_something(self):
-        pass
-
-    def tearDown(self):
-        pass
-
-if __name__ == '__main__':
-    unittest.main()
